added a node for backup the old database
