+++ tms_db/tms_db_manager/CMakeLists.txt
@@ -9,14 +9,14 @@ find_package(catkin REQUIRED COMPONENTS
   tms_msg_db
 )
 
-catkin_python_setup()
-
 catkin_package(
   # INCLUDE_DIRS include
   LIBRARIES tms_db_manager
   CATKIN_DEPENDS roscpp rospy std_msgs geometry_msgs tms_msg_db
   DEPENDS system_lib
 )
+
+catkin_python_setup()
 
 ###########
 ## Build ##
