+++ tms_db/tms_db_manager/setup.py
@@ -1,11 +1,11 @@
-#!/usr/bin/env python
+## ! DO NOT MANUALLY INVOKE THIS setup.py, USE CATKIN INSTEAD
 
 from distutils.core import setup
 from catkin_pkg.python_setup import generate_distutils_setup
 
+# fetch values from package.xml
 setup_args = generate_distutils_setup(
     packages=['tms_db_manager'],
-    package_dir={'': 'src'},
-)
+    package_dir={'': 'src'})
 
 setup(**setup_args)
